therapyGPT/requirements.txt
langchain
openai
python-dotenv
requests
flask
flask-cors
chroma
chromadb
tiktoken